+++ HIMSA Noahlink Wireless Checker.py
@@ -3,8 +3,11 @@
 #                   Copyright Bluebotlabz                   #
 #                                                           #
 #############################################################
+import configparser
 import os
-import feedparser
+import requests
+import lxml.html
+import re
 from urllib.parse import urlparse
 from pathlib import Path
 from colorama import just_fix_windows_console
@@ -42,32 +45,30 @@ disclaimer = [
 if not turboFile.is_file():
     libhearingdownloader.printDisclaimer(disclaimer)
 
-# Get HIMSA Noahlink Wireless Download update using 3rd party web service (PolitePol)
-newsfeed = feedparser.parse('https://politepol.com/fd/smVib0t95riP')
-if(newsfeed.feed == {}):
-    title0 = "NOT FOUND"
+# Read configuration file
+config = configparser.ConfigParser()
+config.read('config.ini')
+uaString = config.get('General', 'UA', fallback='Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/134.0.0.0 Safari/537.3')
+
+# Get HIMSA Noahlink Wireless update from the webpage
+nwURI = "https://www.himsa.com/himsa_download/noahlink-wireless-downloads/?1"
+try:
+    test = requests.get(nwURI, headers={'User-Agent': uaString})
+    dom = lxml.html.fromstring(test.content)
+    hrefs = [x for x in dom.xpath('//a/@href') if '//' in x and 'exe' in x]
+    link0 = hrefs[0].replace('%20', ' ')
+    filename0 = os.path.basename(urlparse(link0).path)
+    title0 = "Noahlink Wireless Firmware Upgrader v" + re.sub(r"\.exe", "", re.sub(r"NLWUpgrader.", "", filename0))
+    link1 = hrefs[1].replace('%20', ' ')
+    filename1 = os.path.basename(urlparse(link1).path)
+    title1 = "Noahlink Wireless Driver v" + re.sub(r"\.exe", "", re.sub(r"Driver_NLW_V.", "", filename1))
+except:
     link0 = ""
     filename0 = "--"
-    title1 = title0
-    link1 = link0
-    filename1 = filename0
-else:
-    try:
-        title0 = newsfeed.entries[0].title
-        link0 = newsfeed.entries[0].link
-        filename0 = os.path.basename(urlparse(link0).path)
-    except:
-        title0 = "NOT FOUND"
-        link0 = ""
-        filename0 = "--"
-    try:
-        title1 = newsfeed.entries[1].title
-        link1 = newsfeed.entries[1].link
-        filename1 = os.path.basename(urlparse(link1).path)
-    except:
-        title1 = "NOT FOUND"
-        link1 = ""
-        filename1 = "--"
+    title0 = "NOT FOUND"
+    link1 = ""
+    filename1 = "--"
+    title1 = "NOT FOUND"
 
 # Define list of valid versions and their download links
 validVersions = [
@@ -78,14 +79,14 @@ validVersions = [
     (" ", "--"),
     ("Archived Downloads", "--"),
     ("==================", "--"),
-    ("Noahlink Wireless Driver v1.1.0.0", "for Windows 10 and 11 (March 2017)", "https://himsafiles.com/NoahlinkWireless/Driver_NLW_V.1.1.0.0.exe"),
     ("Noahlink Wireless Firmware v3.5.0.0", "v2.25 (NW1) / v3.25 (NW2) (July 2025)", "https://www.himsa.com/wp-content/uploads/2025/07/NLWUpgrader_3.5.0.0.exe"),
     ("Noahlink Wireless Firmware v3.3.0.0", "v2.25 (NW1) / v3.23 (NW2) (April 2025)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader_3.3.0.0.exe"),
     ("Noahlink Wireless Firmware v3.1.0.92", "v2.25 (NW1) / v3.17 (NW2) (May 2024)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader_3.1.0.92.exe"),
     ("Noahlink Wireless Firmware v2.24", "v2.24 (NW1 only)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader2.24.exe"),
+    ("Noahlink Wireless Driver v1.1.0.0", "for Windows 10 and 11 (March 2017)", "https://himsafiles.com/NoahlinkWireless/Driver_NLW_V.1.1.0.0.exe"),
 ]
-if(newsfeed.feed == {}):
-    print("\n\nThe latest available version is " + Fore.GREEN + "Firmware v3.3.0.0" + Style.RESET_ALL + " / " + Fore.GREEN + "Driver v1.1.0.0" + Style.RESET_ALL + "\n\n")
+if(link0 == ""):
+    print("\n\nThe latest available version is " + Fore.GREEN + validVersions[7][0] + Style.RESET_ALL + " / " + Fore.GREEN + validVersions[-1][0] + Style.RESET_ALL + "\n\n")
 else:
     print("\n\nThe latest available version is " + Fore.GREEN + title0 + Style.RESET_ALL + " / " + Fore.GREEN + title1 + Style.RESET_ALL + "\n\n")